Remove redundant internal apiConnected flag.
The internal flag to monitor the connection state of the Google API
turned out to be redundant. In rare cases it would not be properly
synced with the API state. This could cause a race condition that
prevents the location updates from being paused when the app is paused,
which would in turn cause unwanted battery drain. Removal of the
redundant flag and use of the API client’s isConnected() fixed the
problem and simplified the code.